Add markdown support and fix build errors.
