+++ frontend/.dockerignore
@@ -1,0 +1,5 @@
+node_modules/
+npm-debug.log
+build
+.dockerignore
+Dockerfile
