+++ frontend/package.json
@@ -6,10 +6,13 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
-    "axios": "^0.27.2"
+    "axios": "^0.27.2",
+    "react-markdown": "^9.0.3",
+    "remark-gfm": "^4.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
