+++ frontend/src/components/ChatControls.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+function ChatControls({ onClear, disabled }) {
+  return (
+    <div className="chat-controls">
+      <button onClick={onClear} className="clear-button" disabled={disabled}>
+        Clear Chat
+      </button>
+    </div>
+  );
+}
+
+export default ChatControls;
